+++ packages/react/src/actions/sendTransaction.ts
@@ -2,6 +2,7 @@ import { VersionedTransaction as SolanaVersionedTransaction } from '@solana/web3
 import { sendTransaction as sendEVMTransaction } from '@wagmi/core';
 import { Address as EVMAddress } from 'viem';
 
+import { Signer, SubmittableExtrinsic } from '@polkadot/api/types';
 import { Cell } from '@ton/core';
 import { CHAIN } from '@tonconnect/ui-react';
 import { ChainData, ChainType, ConnectionOrConfig } from '../types/index.js';
@@ -33,7 +34,7 @@ type TransactionArgs<C extends ChainType> = C extends 'evm'
         }
       : C extends 'alephZero'
         ? {
-            z: string;
+            submittableExtrinsic: SubmittableExtrinsic<'promise' | 'rxjs'>;
           }
         : C extends 'ton'
           ? {
@@ -104,7 +105,41 @@ export const sendTransactionToChain = async <C extends ChainType>({
   }
 
   if (chain.type === 'alephZero') {
+    let txnHash: string | undefined = undefined;
+    let block: string | undefined = undefined;
+    let extrinsicId: number | undefined = undefined;
+
+    const walletConnector = config.connector as WalletInstance<'alephZero'>;
+
     // send transaction to Aleph chain
+    const { submittableExtrinsic } = args as TransactionArgs<'alephZero'>;
+    await submittableExtrinsic.signAndSend(
+      from,
+      { signer: walletConnector.signer as Signer },
+      ({ events, status, txHash, txIndex }) => {
+        events.forEach(({ event }) => {
+          const { method } = event;
+
+          if (method === 'ExtrinsicSuccess' && status.type === 'InBlock') {
+            txnHash = txHash.toString();
+            block = status.asFinalized.toHex();
+            extrinsicId = txIndex;
+          } else if (method === 'ExtrinsicFailed') {
+            throw new Error(`Transaction failed: ${method}`);
+          }
+        });
+      },
+    );
+
+    if (txnHash === undefined || block === undefined) {
+      throw 'Trasaction failed';
+    }
+
+    return {
+      txHash: txnHash,
+      block: block,
+      txIndex: extrinsicId,
+    };
   }
 
   if (chain.type === 'ton') {
@@ -130,6 +165,7 @@ export const sendTransactionToChain = async <C extends ChainType>({
     };
 
     const walletConnector = config.connector as WalletInstance<'ton'>;
+    // send transaction to TON chain
     const tx = await walletConnector.sendTransaction(transaction);
 
     const cell = Cell.fromBase64(tx.boc);