+++ packages/react/src/connectors/evm/walletConfigs.tsx
@@ -13,6 +13,7 @@ export type WalletConfigProps = {
   url?: string;
   icon?: string;
   isInstalled: boolean | (() => boolean) | undefined;
+  hide?: boolean;
   // Create URI for QR code, where uri is encoded data from WalletConnect
   getWalletConnectDeeplink?: (uri: string) => string;
 };
@@ -37,6 +38,7 @@ export const walletConfigs: {
     },
     isInstalled: false,
     icon: '',
+    hide: true,
   },
   coinbaseWalletSDK: {
     name: 'Coinbase Wallet',