+++ packages/react/src/hooks/useConnectionOrConfig.ts
@@ -1,6 +1,7 @@
 import { useConnection as useSolanaConnection } from '@tangled3/solana-react';
 import { useMemo } from 'react';
 import { useConfig as useWagmiConfig } from 'wagmi';
+import { ConnectionOrConfig } from '../types/index.js';
 import { useAlephStore } from './useAlephStore.js';
 import { useTonStore } from './useTonStore.js';
 import { useTronStore } from './useTronStore.js';
@@ -8,7 +9,7 @@ import { useTronStore } from './useTronStore.js';
 /**
  * A hook that returns the connection or config for the all chain types.
  */
-export const useConnectionOrConfig = () => {
+export const useConnectionOrConfig = (): ConnectionOrConfig | undefined => {
   const wagmiConfig = useWagmiConfig();
   const { connection: solanaConnection } = useSolanaConnection();
   const tronWeb = useTronStore((state) => state.tronweb);