+++ packages/react/src/hooks/useWallets.ts
@@ -35,7 +35,7 @@ export const useWallets = (options?: UseWalletsOptions): { [key in ChainType]: W
   const [tonWalletsList, setTonWalletsList] = useState<Wallet<'ton'>[]>([]);
 
   const extendedEvmWallets = useMemo<Wallet<'evm'>[]>(() => {
-    const prepareWallets = (connector: Connector) => {
+    const prepareWallets = (connector: Connector): Wallet<'evm'> | undefined => {
       const walletId = Object.keys(walletConfigs).find(
         (id) =>
           id // where id is comma seperated list
@@ -55,6 +55,7 @@ export const useWallets = (options?: UseWalletsOptions): { [key in ChainType]: W
 
       if (walletId) {
         const wallet = walletConfigs[walletId];
+        if (wallet.hide) return undefined;
 
         return {
           ...c,
@@ -66,7 +67,7 @@ export const useWallets = (options?: UseWalletsOptions): { [key in ChainType]: W
       return c;
     };
 
-    const wallets: Wallet<'evm'>[] = evmConnectors.map((c) => prepareWallets(c));
+    const wallets = evmConnectors.map((c) => prepareWallets(c)).filter((w) => w !== undefined) as Wallet<'evm'>[];
 
     if (options?.onlyInstalled) {
       return wallets.filter((wallet) => wallet.installed);